Add missing git command to pull request guide (#510)
* Add missing git command to pull request guide

The pull request step-by-step guide provides instructions to create a
git branch using upstream/master as the start-point.  This commit adds
a command to the guide which safely makes sure that this start-point
exists in the local repository.  Fixes #509.

* Clarify the need to configure git remote

This added comment explains the need for an existing remote to create a branch from.  The prerequisite git workflow is already documented.

Co-Authored-By: Kyle Stanley <aeros167@gmail.com>

* Move extra git step to its own bullet point